For #24246 - Refactor Tab from HomeFragmentStore
