+++ app/src/main/java/org/mozilla/fenix/collections/CollectionCreationStore.kt
@@ -15,7 +15,6 @@ import mozilla.components.lib.state.State
 import mozilla.components.lib.state.Store
 import org.mozilla.fenix.components.TabCollectionStorage
 import org.mozilla.fenix.ext.toShortUrl
-import org.mozilla.fenix.home.Tab
 
 class CollectionCreationStore(
     initialState: CollectionCreationState
