+++ app/src/main/java/org/mozilla/fenix/collections/Tab.kt
@@ -1,0 +1,16 @@
+/* This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
+
+package org.mozilla.fenix.collections
+
+import android.graphics.Bitmap
+
+data class Tab(
+    val sessionId: String,
+    val url: String,
+    val hostname: String,
+    val title: String,
+    val selected: Boolean? = null,
+    val icon: Bitmap? = null
+)
