+++ app/src/test/java/org/mozilla/fenix/collections/TabDiffUtilTest.kt
@@ -10,7 +10,6 @@ import org.junit.Assert.assertEquals
 import org.junit.Assert.assertFalse
 import org.junit.Assert.assertTrue
 import org.junit.Test
-import org.mozilla.fenix.home.Tab
 
 class TabDiffUtilTest {
